+++ src/login/components/LoginApp/Login/OtherDevice.tsx
@@ -1,12 +1,11 @@
-import { faClock } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { fetchNext, fetchUseOtherDevice, LoginUseOtherResponse } from "apis/eduidLogin";
 import { useAppDispatch, useAppSelector } from "login/app_init/hooks";
 import ButtonPrimary from "login/components/Buttons/ButtonPrimary";
 import React, { useEffect, useState } from "react";
 import { FormattedMessage } from "react-intl";
 import { useHistory, useParams } from "react-router-dom";
-import { TimeRemaining, TimeRemainingWrapper } from "components/TimeRemaining";
+import { TimeRemainingWrapper } from "components/TimeRemaining";
+import { ExpiresMeter } from "./ExpiresMeter";
 
 // URL parameters passed to this component
 interface UseOtherParams {
@@ -111,35 +110,6 @@ function InfoAboutOtherDevice(props: { data: LoginUseOtherResponse }): JSX.Eleme
           </tr>
         </tbody>
       </table>
-    </div>
-  );
-}
-
-interface ExpiresMeterProps {
-  expires_max: number;
-  time_remaining?: TimeRemaining;
-}
-
-export function ExpiresMeter(props: ExpiresMeterProps): JSX.Element | null {
-  // convince TypeScript that TimeRemainingWrapper has added the time_remaining prop
-  if (props.time_remaining === undefined) {
-    return null;
-  }
-  return (
-    <div className="expires-meter">
-      <span className="clock">
-        <FontAwesomeIcon icon={faClock} />
-      </span>
-      <meter
-        low={props.expires_max * 0.2}
-        max={props.expires_max}
-        value={props.time_remaining.total_seconds}
-        id="expires-meter"
-        key="0"
-      />
-      <span className="timer">
-        {props.time_remaining.minutes}:{props.time_remaining.seconds}
-      </span>
     </div>
   );
 }
